Calculated correct amount for transactions and added all outputs to transaction details page

+++ app/src/main/java/com/dcrandroid/data/Transaction.java
@@ -8,8 +8,8 @@ import java.util.ArrayList;
  */
 
 public class Transaction implements Serializable{
-    private float amount, TransactionFee;
-    public float totalInput, totalOutput;
+    private long amount, TransactionFee;
+    public long totalInput, totalOutput;
     private String type, hash, txDate;
     private int height, direction;
     private long time;
@@ -47,11 +47,11 @@ public class Transaction implements Serializable{
         this.txDate = txDate;
     }
 
-    public float getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
 
@@ -59,7 +59,7 @@ public class Transaction implements Serializable{
         return TransactionFee;
     }
 
-    public void setTransactionFee(float transactionFee) {
+    public void setTransactionFee(long transactionFee) {
         this.TransactionFee = transactionFee;
     }
